task(26724548): review & update katas with xfail / not_completed markers (#21)
* task(26724548): update "Python dynamic classes #1" kata

* task(26724548): extend TestCase with test_function arg to use

* task(26724548): update "Pagination helper" solution

+++ katas/5kyu/Pagination helper/solution.py
@@ -57,6 +57,6 @@ class PaginationHelper:
     # this method should return -1 for item_index values that are out of range
     def page_index(self, item_index):
         if self.item_count() > item_index >= 0:
-            return item_index / self._items_per_page
+            return int(item_index / self._items_per_page)
 
         return -1
